Updated notebooks with corrections and bug fixes
